Modified Error

+++ Parser.hpp
@@ -14,8 +14,8 @@ public:
 
     Parser(const std::string & str, const char & delimeter)
     {
-        int prev_index = 0;
-        int deli_index = str.find(delimeter);
+        unsigned long prev_index = 0;
+        unsigned long deli_index = str.find(delimeter);
         while (deli_index != std::string::npos)
         {
             std::string substring = str.substr(prev_index, deli_index - prev_index);
